# openapi.yaml
openapi: 3.0.3
info:
  title: Portfolio Management System API
  description: |
    Multi-client, multi-strategy investment portfolio management backend.

    **Hierarchy:** Client → Account → Sleeve (Account × Strategy join) → Trades/Positions

    - A **Strategy** is a firm-wide definition (e.g. "Growth", "Income") — global and reusable.
    - A **Sleeve** is one strategy as run in one account — it owns trades, positions, and sync configs.
    - All resource IDs are UUIDs (string).
  version: 1.0.0
  contact:
    email: venkatasudheerc@gmail.com
servers:
  - url: http://localhost:8000
    description: Local development server
...
paths:
...
  /api/v1/clients/lookup:
    get:
      tags: [clients]
      summary: Look up a client by email address
      operationId: lookupClientByEmail
      parameters:
        - name: email
          in: query
          required: true
          schema:
            type: string
            format: email
          description: Client email address
      responses:
        "200":
          description: Client found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ClientResponse"
        "404":
          $ref: "#/components/responses/NotFound"
  /api/v1/clients/login:
    post:
      tags: [clients]
      summary: Identify a client by email + dev password
      description: Returns the client record on success. Dev-only — not a production auth mechanism.
      operationId: loginClient
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ClientLogin"
      responses:
        "200":
          description: Login successful — client returned
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ClientResponse"
        "401":
          description: Incorrect password
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "404":
          $ref: "#/components/responses/NotFound"
  /api/v1/clients/:
    post:
      tags: [clients]
      summary: Create a client
      operationId: createClient
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ClientCreate"
      responses:
        "201":
          description: Client created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ClientResponse"
        "422":
          $ref: "#/components/responses/ValidationError"
  /api/v1/clients/{client_id}:
    get:
      tags: [clients]
      summary: Get a client by ID
      operationId: getClient
      parameters:
        - $ref: "#/components/parameters/client_id"
      responses:
        "200":
          description: Client details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ClientResponse"
        "404":
          $ref: "#/components/responses/NotFound"
  /api/v1/clients/{client_id}/performance:
    get:
      tags: [clients]
      summary: Client-level performance roll-up (MWR/TWR)
      description: Merges all trade flows across every sleeve under this client. Returns per-account children.
      operationId: getClientPerformance
      parameters:
        - $ref: "#/components/parameters/client_id"
        - $ref: "#/components/parameters/start_date"
        - $ref: "#/components/parameters/end_date"
      responses:
        "200":
          description: Client performance with account-level children
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LevelPerformance"
        "404":
          $ref: "#/components/responses/NotFound"
  /api/v1/clients/{client_id}/performance/monthly:
    get:
      tags: [clients]
      summary: Client month-on-month returns since inception
      operationId: getClientMonthlyPerformance
      parameters:
        - $ref: "#/components/parameters/client_id"
      responses:
        "200":
          description: Monthly return breakdown
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MonthlyReturnsResponse"
        "404":
          $ref: "#/components/responses/NotFound"
  /api/v1/clients/{client_id}/performance/returns-series:
    get:
      tags: [clients]
      summary: Weekly TWR + MWR timeseries for a client
      operationId: getClientReturnsSeries
      parameters:
        - $ref: "#/components/parameters/client_id"
        - name: start_date
          in: query
          required: false
          schema:
            type: string
            format: date
        - name: end_date
          in: query
          required: false
          schema:
            type: string
            format: date
      responses:
        "200":
          description: Weekly returns series
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ReturnsSeriesResponse"
        "404":
          $ref: "#/components/responses/NotFound"
  /api/v1/clients/{client_id}/positions:
    get:
      tags: [clients]
      summary: Merged positions across all sleeves for a client
      operationId: getClientPositions
      parameters:
        - $ref: "#/components/parameters/client_id"
      responses:
        "200":
          description: Aggregated position summary
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ClientPositionsResponse"
        "404":
          $ref: "#/components/responses/NotFound"
  /api/v1/clients/{client_id}/trades:
    get:
      tags: [clients]
      summary: All trades across all sleeves for a client
      operationId: getClientTrades
      parameters:
        - $ref: "#/components/parameters/client_id"
        - $ref: "#/components/parameters/skip"
        - $ref: "#/components/parameters/limit"
      responses:
        "200":
          description: Paginated trade list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ClientTradesResponse"
        "404":
          $ref: "#/components/responses/NotFound"

  # ─── Accounts ───────────────────────────────────────────────────────────────
...
components:
  parameters:
    client_id:
      name: client_id
      in: path
      required: true
      schema:
        type: string
      description: Client UUID
...
    start_date:
      name: start_date
      in: query
      required: false
      schema:
        type: string
        format: date
      description: Filter start date (YYYY-MM-DD)
    end_date:
      name: end_date
      in: query
      required: false
      schema:
        type: string
        format: date
      description: Filter end date (YYYY-MM-DD)
    skip:
      name: skip
      in: query
      required: false
      schema:
        type: integer
        minimum: 0
        default: 0
      description: Pagination offset
    limit:
      name: limit
      in: query
      required: false
      schema:
        type: integer
        minimum: 1
        maximum: 1000
        default: 100
      description: Maximum number of results to return
  responses:
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
...
    ValidationError:
      description: Request body failed Pydantic validation
      content:
        application/json:
          schema:
            type: object
            properties:
              detail:
                type: array
                items:
                  type: object
                  properties:
                    loc:
                      type: array
                      items: {}
                    msg:
                      type: string
                    type:
                      type: string
  schemas:
    Error:
      type: object
      properties:
        detail:
          type: string
      required: [detail]

    # ── Clients ──────────────────────────────────────────────────────────────
    ClientLogin:
      type: object
      required: [email, password]
      properties:
        email:
          type: string
          format: email
        password:
          type: string
    ClientCreate:
      type: object
      required: [name, email]
      properties:
        name:
          type: string
          example: Acme Corp
        email:
          type: string
          format: email
          example: contact@acme.com
    ClientResponse:
      type: object
      properties:
        id:
          type: string
          example: 550e8400-e29b-41d4-a716-446655440000
        name:
          type: string
        email:
          type: string
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    ClientPositionsResponse:
      type: object
      properties:
        client_id:
          type: string
        total_cost_basis:
          type: number
          format: float
        total_market_value:
          type: number
          format: float
        total_unrealized_gain:
          type: number
          format: float
        positions_count:
          type: integer
        positions:
          type: array
          items:
            $ref: "#/components/schemas/AggregatedPosition"
    AggregatedPosition:
      type: object
      properties:
        symbol:
          type: string
        quantity:
          type: number
        cost_basis:
          type: number
        avg_cost:
          type: number
        current_price:
          type: number
          nullable: true
        market_value:
          type: number
          nullable: true
        unrealized_gain:
          type: number
          nullable: true
        unrealized_gain_pct:
          type: number
          nullable: true
        trades_count:
          type: integer
    ClientTradesResponse:
      type: object
      properties:
        client_id:
          type: string
        total:
          type: integer
        skip:
          type: integer
        limit:
          type: integer
        trades:
          type: array
          items:
            $ref: "#/components/schemas/TradeSummary"
    TradeSummary:
      type: object
      properties:
        id:
          type: string
        sleeve_id:
          type: string
        trade_date:
          type: string
          format: date
        symbol:
          type: string
        action:
          $ref: "#/components/schemas/TradeAction"
        quantity:
          type: number
        price:
          type: number
        commission:
          type: number
        notes:
          type: string

    # ── Accounts ─────────────────────────────────────────────────────────────
...
    TradeAction:
      type: string
      enum: [BUY, SELL]
...
    PerformancePoint:
      type: object
      properties:
        date:
          type: string
          format: date
        value:
          type: number
          nullable: true
    PerformanceSummary:
      type: object
      properties:
        mwr_pct:
          type: number
          nullable: true
          description: Money Weighted Return (IRR) as a decimal (e.g. 0.12 = 12%)
        twr_pct:
          type: number
          nullable: true
          description: Time Weighted Return (geometric linking) as a decimal
        twr_note:
          type: string
          nullable: true
        total_return_pct:
          type: number
          nullable: true
        total_cost_basis:
          type: number
        total_market_value:
          type: number
        total_unrealized_gain:
          type: number
        total_invested:
          type: number
        total_proceeds:
          type: number
        total_realized_gain:
          type: number
        closed_positions_count:
          type: integer
        trades_count:
          type: integer
    ChildPerformance:
      type: object
      properties:
        level:
          type: string
        id:
          type: string
        name:
          type: string
          nullable: true
        summary:
          $ref: "#/components/schemas/PerformanceSummary"
    LevelPerformance:
      type: object
      properties:
        level:
          type: string
          enum: [client, account, sleeve, strategy]
        id:
          type: string
        name:
          type: string
        start_date:
          type: string
          format: date
          nullable: true
        end_date:
          type: string
          format: date
          nullable: true
        summary:
          $ref: "#/components/schemas/PerformanceSummary"
        timeseries:
          type: array
          items:
            $ref: "#/components/schemas/PerformancePoint"
        children:
          type: array
          items:
            $ref: "#/components/schemas/ChildPerformance"
    MonthlyReturn:
      type: object
      properties:
        year:
          type: integer
        month:
          type: integer
        month_label:
          type: string
        start_value:
          type: number
        end_value:
          type: number
        cash_in:
          type: number
        cash_out:
          type: number
        net_cash_flow:
          type: number
        realized_gain:
          type: number
        return_pct:
          type: number
          nullable: true
          description: Monthly return as a decimal (e.g. 0.034 = 3.4%); null for inception month
    MonthlyReturnsResponse:
      type: object
      properties:
        months:
          type: array
          items:
            $ref: "#/components/schemas/MonthlyReturn"
        cumulative_return_pct:
          type: number
          nullable: true

    # ── Weekly Returns Series ────────────────────────────────────────────────
    WeeklyReturnPoint:
      type: object
      properties:
        date:
          type: string
          format: date
          description: Monday of the snapshot week
        twr_cumul:
          type: number
          nullable: true
          description: Cumulative TWR from inception to this week (decimal, e.g. 0.05 = 5%)
        mwr_cumul:
          type: number
          nullable: true
          description: Cumulative MWR (IRR holding-period) from inception to this week
    ReturnsSeriesResponse:
      type: object
      properties:
        level:
          type: string
          enum: [client, account, sleeve, strategy]
        id:
          type: string
        series:
          type: array
          items:
            $ref: "#/components/schemas/WeeklyReturnPoint"

    # ── Market Prices ────────────────────────────────────────────────────────
...